extract components into a constants file
